Make `DynMessage` a struct, use trait upcasting (bump MSRV to 1.86) (#920)
With trait upcasting stabilised in Rust 1.86, we can make use of it in
`DynMessage`.

This also makes `DynMessage` into a struct, because that reduces the
risk of Rust Analyzer messing things up; it liked to confuse the
`Message` trait with the `Message` generic. That is, the prior state was
effectively `trait View<Message = Box<dyn Message>>`, which sometimes
led to things like: `Box<dyn Box<dyn Message>>`.