Xilem: Pass `app_state` as a parameter to `View::{build, rebuild, teardown}` (#1079)
This change is both a major departure from the prior state, and the most
natural thing in the world.

My proposing this change has come from thinking related to #1070, for
what build/rebuild/message actually represent, , how distinct important
the phases of the lifecycle actually are (they aren't at all), and the
reason that the Xilem model actually works. The core insight behind
Xilem is that if you can prove to the compiler that you only access a
piece of state in one place at once, the compiler will let you access it
exclusively. Everything else is following that to its logical
conclusion; this just continues that train of thought.

Current state of this PR:
- This PR is ready to be merged
- This PR makes a lot of changes, so if we want this, merging quickly
would be helpful
- The only thing which is broken is the `adapt` view, which currently
unconditionally panicks, to avoid making behaviour changes in this PR
- The follow-up changes which make use of this (in task, worker,
virtual_scroll, memoize, the new use_state, run_once, etc.) have not
been done, to make this PR plausible to review.

Tradeoffs:
- In theory, this makes multi-threaded rebuild harder, but note that
this was already probably hard
- This is rejecting the premise that you don't need component local
state; this is something I treated as an experiment, but which has
probably held us back significantly.

There's some more work to be done here (in particular, I really want to
experiment with exclusive references in the app state, especially around
`use_state`, for event handling).

---------

Co-authored-by: Philipp Mildenberger <philipp@mildenberger.me>