Migrate Xilem Core and Masonry to Anymore (#1170)
[Anymore](https://github.com/linebender/anymore) is a new crate which
we're developing for use in Xilem and Masonry. This crate contains the
`AnyDebug` trait, which is a supertrait of both Any and Debug. We are
developing this because Xilem Core and Masonry both want support for
dynamically typed messages with introspection for debugging, but those
have to use the same trait to allow proper interoperability.

See also [#rust ui >
Anymore](https://xi.zulipchat.com/#narrow/channel/422907-rust-ui/topic/Anymore/with/528892562)

This crate has not yet published to crates.io; I intend for us to do so
once we have migrated Masonry's `Action` to being only a wrapper around
`Box<dyn AnyDebug + Send>`. See also discussion in #1105